One of the most common uses of ProxyLookup is to dynamically change the set of Lookups being proxies (a quick grep finds 53 such cases in the NetBeans source base, and I can think of > 10 I have written in external modules.  For each consumer of the API that needs to do this, it is necessary to write identical subclasses with a public or package visibility method that exposes one of ProxyLookup's setLookups() methods.
This patch adds ProxyLookup.create(Consumer<Consumer<Lookup[]>>) and
ProxyLookup.create(Consumer<BiConsumer<Executor, Lookup[]>>) to satisfy the
non-threaded and threaded-event-delivery variants supported as protected methods on
ProxyLookup.  These allow a mechanism for updating the set of proxied lookups to
be passed to the caller without exposing that machanism to consumers of the returned
Lookup or any code other that which immediately instantiates it.